refactor: change app directory structure
