+++ App.tsx
@@ -5,20 +5,20 @@ import WelcomeWalkThrough from '@/app/screens/WelcomeWalkThrough';
 import OnboardingStepOne from '@/app/screens/OnboardingStepOne';
 import OnboardingStepTwo from '@/app/screens/OnboardingStepTwo';
 import OnboardingStepThree from '@/app/screens/OnboardScreenThree';
-import CreateNewWallet from '@/app/screens/CreateNewWallet';
+import CreateNewWallet from 'app/CreateNewWallet';
 import ImportExistingWallet from '@/app/screens/ImportExistingWallet';
 import Home from '@/app/screens/Home';
 
 
 export default function App() {
   return (
-    // <WelcomeSplash/>
+    <WelcomeSplash/>
     // <WelcomeWalkThrough/>
     // <OnboardingStepOne/>
     // <OnboardingStepTwo/>
     // <OnboardingStepThree/>
     // <CreateNewWallet/>
     // <ImportExistingWallet/>
-    <Home/>
+    // <Home/>
   );
 }
