+++ app/CreateNewWallet.tsx
@@ -1,10 +1,10 @@
 import { Pressable, StyleSheet, Text, TextInput, View } from 'react-native'
 import React from 'react'
-import { COLORS } from '@/app/COLORS'
+import { COLORS } from 'app/COLORS'
 import { useFonts } from '@expo-google-fonts/poppins/useFonts'
 import { Poppins_600SemiBold, Poppins_700Bold } from '@expo-google-fonts/poppins'
 
-const ImportExistingWallet = () => {
+const CreateNewWallet = () => {
     const [fontsLoaded] = useFonts({
         Poppins_600SemiBold,
         Poppins_700Bold
@@ -14,27 +14,22 @@ const ImportExistingWallet = () => {
         <View style={styles.createWalletContainer}>
             <View style={styles.walletTextContainer}>
                 <Text style={styles.walletContainerHeading}>
-                    Import
-                    Existing Wallet
+                    Create New Wallet
                 </Text>
                 <Text style={styles.walletText}>
-                    Seamlessly transition to Bitgesell Wallet by pasting your mnemonic phrase – effortlessly import your existing wallet and dive into decentralized convenience.
+                    Create a new  Bitgesell Wallet to  manage all your
+                    BGL and other assets
                 </Text>
             </View>
             <View style={styles.createWalletForm}>
                 <Text style={styles.inputLabel}>
-                    Seedphrase/Mnemonic
+                    Enter Password
                 </Text>
-                <TextInput style={styles.passwordInput} placeholder='Paste wallet seed phrase' placeholderTextColor={COLORS.WHITE003} />
-                <Text style={styles.inputLabel}>
-                    Password
-                </Text>
-                <TextInput style={styles.passwordInput} placeholder='Password' placeholderTextColor={COLORS.WHITE003} />
-
+                <TextInput style={styles.passwordInput} textContentType='password' />
                 <Text style={styles.walletText}>
                     🔒 Keep this password safe as it will be used to keep your wallet safe on this device.
                 </Text>
-                <View>
+                <View style={styles.submitButtonContainer}>
                     <Pressable style={styles.submitButton} onPress={() => alert('Im in!')}>
                         <Text style={styles.submitButtonText}>Create a A New Wallet</Text>
                     </Pressable>
@@ -44,7 +39,7 @@ const ImportExistingWallet = () => {
     )
 }
 
-export default ImportExistingWallet
+export default CreateNewWallet
 
 const styles = StyleSheet.create({
     createWalletContainer: {
@@ -113,7 +108,7 @@ const styles = StyleSheet.create({
         color: COLORS.WHITE,
         fontSize: 16
     },
-    // submitButtonContainer: {
-    //     marginTop: 158
-    // }
+    submitButtonContainer: {
+        marginTop: 158
+    }
 })
