+++ components/IsometricOne.tsx
