+++ components/SliderDots.tsx
