+++ components/SplashLogo.tsx
