+++ tsconfig.json
@@ -4,7 +4,7 @@
     "strict": true,
     "baseUrl": "./",
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./*"]
     }  
   },
 }
